packages/backend-django/requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
dotenv==0.9.9
idna==3.10
iniconfig==2.1.0
Markdown==3.8
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
